Move device link to services namespace and make checking more thourough
